requirements.txt
bleach>=6.2,<7
django>=5.2,<5.3
django-cors-headers>=4.7,<5
django-csp>=4.0,<5
django-decouple>=3.8,<4
django-ratelimit>=4.1,<5
djangorestframework>=3.16,<4
djangorestframework-simplejwt>=5.5,<6
pillow>=11.2,<12
qrcode>=8.2,<9
requests>=2.32,<3
bleach>=6.2,<7
Django>=5.2,<5.3
django-cors-headers>=4.7,<5
django-csp>=4.0,<5
django-ratelimit>=4.1,<5
djangorestframework>=3.16,<4
djangorestframework-simplejwt>=5.5,<6
Pillow>=11.0,<12
psycopg2-binary>=2.9,<3
pyjwt>=2.10,<3
python-decouple>=3.8,<4
qrcode>=8.2,<9
pytest>=8.3,<9
pytest-django>=4.11,<5
requests>=2.32,<3
